requirements.txt
flask==3.0.0
gunicorn==21.2.0
tensorflow==2.19.0
keras==3.10.0
numpy==1.26.4
pillow==10.3.0
pandas==2.2.2
h5py==3.11.0